initialize frontend of tch
