--- Downloads/To_Do_App-main/src/Components/index.js
@@ -1,4 +1,0 @@
-import TodoForm from "./TodoForm"
-import TodoItem from "./TodoItem"
-
-export { TodoForm, TodoItem }
