--- Downloads/To_Do_App-main/src/TodoContexts/TodoContext.js
@@ -1,22 +1,0 @@
-import {createContext, useContext} from "react"
-
-export const TodoContext = createContext({
-    todos: [
-        {
-            id: 1,
-            todo: " Todo msg",
-            completed: false,
-        }
-    ],
-    addTodo: (todo) => {},
-    updateTodo: (id, todo) => {},
-    deleteTodo: (id) => {},
-    toggleComplete: (id) => {}
-})
-
-
-export const useTodo = () => {
-    return useContext(TodoContext)
-}
-
-export const TodoProvider = TodoContext.Provider
