--- Downloads/To_Do_App-main/tailwind.config.js
@@ -1,10 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-export default {
-  content: ["./index.html",
-    "./src/**/*.{js,ts,jsx,tsx}",],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
-
