--- OneDrive/Desktop/acehack/aiweb_generator/frontend/src/App.tsx
@@ -1,13 +1,0 @@
-import './App.css'
-
-function App() {
-  
-
-  return (
-    <>
-      <h1 className='text-3xl text-red-500'>hello</h1>
-    </>
-  )
-}
-
-export default App
