--- OneDrive/Desktop/acehack/aiweb_generator/frontend/vite.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from 'vite'
-import tailwindcss from '@tailwindcss/vite'
-export default defineConfig({
-  plugins: [
-    tailwindcss(),
-  ],
-})
-
