+++ src/App.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function App() {
+  return (
+    <div className='bg-red-500 text-amber-300 text-2xl font-extrabold '>TCH Frontend</div>
+  )
+}
+
+export default App
